Add files via upload
This is a subset from the 2020 National Survey on Drug Use and Health, filtered to include only respondents under 18. The data was further subset to include 79 variables relating to substance use, family and home life, demographics, and education. These were chosen from the original 2,892 variables in order to improve model simplicity and reduce computational costs.